+++ src/config/default.js
@@ -68,6 +68,7 @@ export default {
     findAllCharacters: hours(1),
     statistics: hours(1),
     resourceOfTheDay: hours(24),
+    gw2Api: minutes(5),
   },
 
   pagination: {